+++ Changelog.md
@@ -8,6 +8,7 @@ Features:
  * Inline Assembly: Issue warning for using jump labels (already existed for jump instructions).
  * Inline Assembly: Support some restricted tokens (return, byte, address) as identifiers in Julia mode.
  * Optimiser: Replace `x % 2**i` by `x & (2**i-1)`.
+ * Resolver: Suggest alternative identifiers if a given identifier is not found.
  * SMT Checker: If-else branch conditions are taken into account in the SMT encoding of the program
    variables.
  * Syntax Checker: Deprecate the ``var`` keyword (and mark it an error as experimental 0.5.0 feature).